move submissions into separate json with method injection
