+++ src/v2/firestorm/packs/submissions.ts
@@ -1,0 +1,7 @@
+import firestorm from "firestorm-db";
+import config from "../config";
+import { FirestormSubmission } from "~/v2/interfaces";
+
+config();
+
+export const submissions = firestorm.collection<FirestormSubmission>("submissions");
