+++ src/v2/interfaces/index.ts
@@ -13,3 +13,4 @@ export * from "./textures";
 export * from "./users";
 export * from "./uses";
 export * from "./packs";
+export * from "./submissions";
