+++ src/v2/repository/firestorm/packs.repository.ts
@@ -1,6 +1,6 @@
 import { ID_FIELD } from "firestorm-db";
 import { PackRepository, Pack, PackTag, Packs } from "~/v2/interfaces";
-import { packs } from "../../firestorm/settings/packs";
+import { packs } from "../../firestorm/packs";
 
 export default class PackFirestormRepository implements PackRepository {
 	getRaw(): Promise<Record<string, Pack>> {
